Move AssemblyResolver to separate file.
