+++ src/EditorFeatures/CSharp/DecompiledSource/CSharpDecompiledSourceService.cs
@@ -71,14 +71,13 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.DecompiledSource
             return document;
         }
 
-        static Document PerformDecompilation(Document document, string fullName, Compilation compilation, string assemblyLocation)
+        private static Document PerformDecompilation(Document document, string fullName, Compilation compilation, string assemblyLocation)
         {
             // Load the assembly.
             var pefile = new PEFile(assemblyLocation, PEStreamOptions.PrefetchEntireImage);
 
             // Initialize a decompiler with default settings.
-            var settings = new DecompilerSettings(LanguageVersion.Latest);
-            var decompiler = new CSharpDecompiler(pefile, new RoslynAssemblyResolver(compilation), settings);
+            var decompiler = new CSharpDecompiler(pefile, new AssemblyResolver(compilation), new DecompilerSettings());
             // Escape invalid identifiers to prevent Roslyn from failing to parse the generated code.
             // (This happens for example, when there is compiler-generated code that is not yet recognized/transformed by the decompiler.)
             decompiler.AstTransforms.Add(new EscapeInvalidIdentifiers());
@@ -99,62 +98,6 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.DecompiledSource
             return document.WithText(SourceText.From(header + text));
         }
 
-        private class RoslynAssemblyResolver : IAssemblyResolver
-        {
-            private readonly Compilation parentCompilation;
-            private static readonly Version zeroVersion = new Version(0, 0, 0, 0);
-
-            public RoslynAssemblyResolver(Compilation parentCompilation)
-            {
-                this.parentCompilation = parentCompilation;
-            }
-
-            public PEFile Resolve(IAssemblyReference name)
-            {
-                foreach (var assembly in parentCompilation.GetReferencedAssemblySymbols())
-                {
-                    // First, find the correct IAssemblySymbol by name and PublicKeyToken.
-                    if (assembly.Identity.Name != name.Name
-                        || !assembly.Identity.PublicKeyToken.SequenceEqual(name.PublicKeyToken ?? Array.Empty<byte>()))
-                    {
-                        continue;
-                    }
-
-                    // Normally we skip versions that do not match, except if the reference is "mscorlib" (see comments below)
-                    // or if the name.Version is '0.0.0.0'. This is because we require the metadata of all transitive references
-                    // and modules, to achieve best decompilation results.
-                    // In the case of .NET Standard projects for example, the 'netstandard' reference contains no references
-                    // with actual versions. All versions are '0.0.0.0', therefore we have to ignore those version numbers,
-                    // and can just use the references provided by Roslyn instead.
-                    if (assembly.Identity.Version != name.Version && name.Version != zeroVersion
-                        && !string.Equals("mscorlib", assembly.Identity.Name, StringComparison.OrdinalIgnoreCase))
-                    {
-                        // MSBuild treats mscorlib special for the purpose of assembly resolution/unification, where all
-                        // versions of the assembly are considered equal. The same policy is adopted here.
-                        continue;
-                    }
-
-                    // reference assemblies should be fine here, we only need the metadata of references.
-                    var reference = parentCompilation.GetMetadataReference(assembly);
-                    return new PEFile(reference.Display, PEStreamOptions.PrefetchMetadata);
-                }
-
-                // not found
-                return null;
-            }
-
-            public PEFile ResolveModule(PEFile mainModule, string moduleName)
-            {
-                // Primitive implementation to support multi-module assemblies
-                // where all modules are located next to the main module.
-                string baseDirectory = Path.GetDirectoryName(mainModule.FileName);
-                string moduleFileName = Path.Combine(baseDirectory, moduleName);
-                if (!File.Exists(moduleFileName))
-                    return null;
-                return new PEFile(moduleFileName, PEStreamOptions.PrefetchMetadata);
-            }
-        }
-
         private string GetFullReflectionName(INamedTypeSymbol containingType)
         {
             var stack = new Stack<string>();
@@ -169,6 +112,5 @@ namespace Microsoft.CodeAnalysis.Editor.CSharp.DecompiledSource
 
             return string.Join(".", stack);
         }
-
     }
 }
